Extract common foreign chain call interface, use it
