+++ Cargo.toml
@@ -11,6 +11,7 @@ members = [
 [workspace.dependencies]
 anyhow = "1.0.65"
 assert_matches = "1.5.0"
+cosmwasm-std = "1.0.0"
 ethabi = "17.2.0"
 eyre = "0.6.8"
 hex = "0.4.3"
@@ -21,6 +22,7 @@ serde = { version = "1.0.140", default-features = false, features = ["derive"] }
 serde_json = "1.0.85"
 sha3 = "0.10.5"
 thiserror = "1.0.35"
+xcci = { path = "packages/xcci" }
 
 [profile.release]
 opt-level = "z"
