+++ egg/mint/src/msg.rs
@@ -1,6 +1,7 @@
-use cosmwasm_std::{Binary, CustomMsg};
+use cosmwasm_std::Binary;
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
+use xcci::TargetContractInfo;
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
 pub struct MigrateMsg {}
@@ -8,14 +9,6 @@ pub struct MigrateMsg {}
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
 pub struct InstantiateMsg {
     pub target_contract_info: TargetContractInfo,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
-pub struct TargetContractInfo {
-    pub chain_id: String,
-    pub compass_id: String,
-    pub contract_address: String,
-    pub smart_contract_abi: String,
 }
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
@@ -28,12 +21,3 @@ pub enum ExecuteMsg {
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
 #[serde(rename_all = "snake_case")]
 pub enum QueryMsg {}
-
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
-#[serde(rename_all = "snake_case")]
-pub struct CustomResponseMsg {
-    pub target_contract_info: TargetContractInfo,
-    pub payload: Binary,
-}
-
-impl CustomMsg for CustomResponseMsg {}
