+++ egg/mint/src/state.rs
@@ -1,7 +1,7 @@
-use crate::msg::TargetContractInfo;
 use cosmwasm_std::Addr;
 use cw_storage_plus::{Item, Map};
 use std::collections::HashSet;
+use xcci::TargetContractInfo;
 
 pub const ADMIN: Item<Addr> = Item::new("admin");
