+++ egg/robin/src/msg.rs
@@ -1,20 +1,13 @@
-use cosmwasm_std::{Binary, CustomMsg};
+use cosmwasm_std::Binary;
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
+use xcci::TargetContractInfo;
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
 pub struct MigrateMsg {}
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
 pub struct InstantiateMsg {}
-
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
-pub struct TargetContractInfo {
-    pub chain_id: String,
-    pub contract_address: String,
-    pub smart_contract_abi: String,
-    pub compass_id: String,
-}
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
 #[serde(rename_all = "snake_case")]
@@ -24,15 +17,6 @@ pub enum ExecuteMsg {
         payload: Binary,
     },
 }
-
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
-#[serde(rename_all = "snake_case")]
-pub struct CustomResponseMsg {
-    pub target_contract_info: TargetContractInfo,
-    pub payload: Binary,
-}
-
-impl CustomMsg for CustomResponseMsg {}
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
 #[serde(rename_all = "snake_case")]
